Provides its-support and building backends
I thought it may be similar to the existing inspe
resources to use the its syntax. I probably should require
the its hiding inside of inspec.

Made it so that the environment blocks will build on one another but it
would be good to find a way to handle de-dupe or ordering of commands
given the potential there with conflicts.